ANDROID: Revert: Merge 5.4.60 into android11-5.4
Something is going flaky with the ABI checker, so back this merge out at
the moment so that we can fix things up first before merging this back
in.

Bug: 161946584

+++ fs/ocfs2/suballoc.c
@@ -879,9 +879,9 @@ static void __ocfs2_set_steal_slot(struct ocfs2_super *osb, int slot, int type)
 {
 	spin_lock(&osb->osb_lock);
 	if (type == INODE_ALLOC_SYSTEM_INODE)
-		osb->s_inode_steal_slot = (u16)slot;
+		osb->s_inode_steal_slot = slot;
 	else if (type == EXTENT_ALLOC_SYSTEM_INODE)
-		osb->s_meta_steal_slot = (u16)slot;
+		osb->s_meta_steal_slot = slot;
 	spin_unlock(&osb->osb_lock);
 }
